roadiniServer/roadini/requirements.txt
beautifulsoup4==4.4.1
blinker==1.3
Brlapi==0.6.4
chardet==2.3.0
checkbox-support==0.22
defer==1.0.6
Django==2.1.2
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
feedparser==5.1.3
guacamole==0.9.2
html5lib==0.999
httplib2==0.9.1
Jinja2==2.8
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
MarkupSafe==0.23
netifaces==0.10.4
oauthlib==1.0.3
onboard==1.2.0
padme==1.1.1
pexpect==4.0.1
Pillow==5.3.0
plainbox==0.25
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.5
pyasn1==0.1.9
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
PyJWT==1.6.4
pyparsing==2.0.3
python-apt==1.1.0b1+ubuntu0.16.4.2
python-debian==0.1.27
python-systemd==231
pytz==2018.5
pyxdg==0.25
pyzmq==15.2.0
reportlab==3.3.0
requests==2.9.1
six==1.10.0
tornado==4.2.1
ufw==0.35
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
unity-tweak-tool==0.0.7
urllib3==1.13.1
usb-creator==0.3.0
virtualenv==15.0.1
ws4py==0.3.4
xkit==0.0.0
XlsxWriter==0.7.3
